+++ TEST 3 LUX/FORMS/Actividades/DibujarActividades.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace TEST_3_LUX
+{
+    public partial class DibujarActividades : Form
+    {
+        public DibujarActividades()
+        {
+            InitializeComponent();
+        }
+    }
+}
